Merge classic and core gadgets
This is not yet complete, but is about 90% there. Specifically, psplash
(for core) is still being built manually from some commit in the yocto
project. This ought to be packaged, or replaced with something that's
already packaged. A 64-bit boot config for core is also still required.

Finally, the core boot environment should be merged into the classic
boot script, but this requires changes to snapd itself.