Remove redundant patches, fix fdt_high
The uboot patches are no longer required (all u-boot images come from
the u-boot-rpi package). Fixed an issue with uboot re-locating the fdt
on the Pi4; the loaded environment was wiping the fdt_high and
initrd_high values in the environment. These are now restored as early
as possible in bootcmd.