Fix dependency for boot-script target
The device-trees dependency should be against boot-script rather than
config-classic. It's required to calculate the latest kernel version (in
a slightly hacky manner with a for-loop), in order to generate
substitution values for the boot script. Originally this was against the
boot-classic target as that's the only one that used the boot script but
now that both core and classic use the boot script, the dependency
should be against the boot-script target instead.