Makefile, snapcraft.yaml: specify Dir::Etc::sourceparts too
In recent focal builder images that one would use to build this snap locally,
when cross-compiling there are now files in this directory in the build
environment which we want to ignore when downloading manually with our
stage-package helper in the Makefile. However, when natively compiling we will
want to copy these sources from the host in case the snappy-dev image PPA was
installed on the host, so that we can still build using the correct sources
here.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>