Disable LED customization; tidy up config.txt
Prior versions of Core customized the ACT & PWR LEDs. This commit
disables those customizations but leaves a comment to trivially
re-enable them if users should desire. It also adds some commentary to
the other settings in a similar style to that used in the classic
config.txt.

Once merged, this should be cherry-picked to 20-armhf (and possibly
classic & desktop, though only to minimize the branch diffs).

Signed-off-by: Dave Jones <dave.jones@canonical.com>