Switch gadget to use new `$kernel:ref` dtb handling
Snapd is in the process of supporting DTBs coming from the kernel
snap. This requires an update of the gadget.yaml to point to the
kernel snap for boot assets of the form `$kernel:ref`.

The coresponding kernel snap change requires adding a file
meta/kernel.yaml with the content:
```
assets:
  pidtbs:
    update: true
    content:
    - dtbs/broadcom/
    - dtbs/overlays/
```

This can only land for real once snapd is fully ready, the last
PR that needs to land is https://github.com/snapcore/snapd/pull/9999

Note that this PR also tweaks the "classic" target as well but I did
not test this, because I don't know how it's called. For classic
the new syntax cannot be used because ubuntu-image needs help from
`snap prepare-image` to do the resolving of the content. But the
new `no-kernel-refs-gadget` in the makefile should ensure things
work like before for classic images.